correccion en funcion recuperar para ultimo archivo

+++ APL1/APL1_EJ6/Ejercicio6.sh
@@ -103,15 +103,15 @@ recuperar(){
         nombreArchivo=${nombre% (copia*}    # mejora para archivo mismo nombre, mismo directorio
         if [ "$nombreArchivo" == "$archivoParaRecuperar" ];
         then
-	   (( contadorArchivosIguales++ ))
-	    rutaArchivoUnico="$archivo"
-           if [ "$contadorArchivosIguales" -lt 1 ]
-	   then
-		break;
-	   fi
-	fi
+	        (( contadorArchivosIguales++ ))
+	        rutaArchivoUnico="$archivo"
+            if [ "$contadorArchivosIguales" -lt 1 ]
+	        then
+		        break;
+	        fi
+	    fi
     done
-    #echo "cantidad: $contadorArchivosIguales"
+    
     if [ "$contadorArchivosIguales" -eq 0 ];
     then
         echo "No existe el archivo en la papelera"
@@ -119,32 +119,34 @@ recuperar(){
     else
         if [ "$contadorArchivosIguales" -eq 1 ];	#Archivo unico
         then
-            tar -Pxvf "$papelera" "$rutaArchivoUnico" 1> /dev/null
+            rutaArchivo=$(dirname "$rutaArchivoUnico")
+            nombre=$(basename "$rutaArchivoUnico")
+            nombreArchivo=${nombre% (*}
+            archivoFinal="${rutaArchivo}/${nombreArchivo}"
+            tar -Pxvf "$papelera" "$rutaArchivoUnico" -O > "$archivoFinal" 1> /dev/null
             tar -Pvf "$papelera" --delete "$rutaArchivoUnico" 1> /dev/null
         else
 		#Mas de un archivo con el mismo nombre
 
-	   contadorArchivosIguales=0
-	   archivosIguales=0
-	   archivosIguales=""
-           IFS=$'\n'
-   	   for archivo in $(realpath $(tar -Ptf "$papelera"))
-    	   do
-        		rutaArchivo=$(dirname "$archivo")
-        		nombre=$(basename "$archivo")
+	        contadorArchivosIguales=0
+	        archivosIguales=0
+	        archivosIguales=""
+            IFS=$'\n'
+   	        for archivo in $(realpath $(tar -Ptf "$papelera"))
+    	    do
+                rutaArchivo=$(dirname "$archivo")
+            	nombre=$(basename "$archivo")
                 nombreArchivo=${nombre% (copia*}    # mejora para archivo mismo nombre, mismo directorio
-                #echo "rutaArchivo: $rutaArchivo"
-                #echo "nombreArchivo: $nombreArchivo"
-        	if [ "$nombreArchivo" == "$archivoParaRecuperar" ];
-        	then
-            		let contadorArchivosIguales=contadorArchivosIguales+1
-            		#archivosIguales="$archivosIguales$contadorArchivosIguales - $nombreArchivo $rutaArchivo;"
-                    archivosIguales="$archivosIguales$contadorArchivosIguales - $nombre $rutaArchivo;"
-            		arrayArchivos[$contadorArchivosIguales]="$archivo"
-        	fi
-   	    done
+                
+            	if [ "$nombreArchivo" == "$archivoParaRecuperar" ];
+            	then
+                	let contadorArchivosIguales=contadorArchivosIguales+1
+                	archivosIguales="$archivosIguales$contadorArchivosIguales - $nombre $rutaArchivo;"
+                	arrayArchivos[$contadorArchivosIguales]="$archivo"
+            	fi
+   	        done
 
-	    echo "$archivosIguales" | awk 'BEGIN{FS=";"} {for(i=1; i < NF; i++) print $i}'
+	        echo "$archivosIguales" | awk 'BEGIN{FS=";"} {for(i=1; i < NF; i++) print $i}'
             echo "¿Qué archivo desea recuperar?"
             read opcion
 
@@ -166,7 +168,6 @@ recuperar(){
             IFS=$'\n'
             for archivo in $(tar -Ptf "$papelera" ) 
             do  
-
                 let indice=$indice+1
                 if [ "$indice" == "$elementoNumero" ];
                 then
@@ -185,8 +186,11 @@ recuperar(){
 
 vaciar(){
     papelera="${HOME}/papelera.zip"
-    rm "$papelera"
-    tar -Pcf "$papelera" --files-from /dev/null
+    if [ -f "$papelera" ];
+    then
+        rm "$papelera"
+        tar -Pcf "$papelera" --files-from /dev/null
+    fi
     echo "Se elimino definitivamente todos los archivos"
 }
 
@@ -223,7 +227,6 @@ eliminar(){
                 fi
             fi
         done
-        #echo "eliminar-rutaArchivoUnico: $rutaArchivoUnico"
 
         # renombrar archivo con numero de secuencia
         if [ "$contadorArchivosIguales" -ne 0 ];
@@ -238,7 +241,6 @@ eliminar(){
                 (( repetidosss++ ))
                 nombreCopia="$archivoEliminar (copia $repetidosss)"
             fi
-            #nombreCopia="$archivoEliminar (copia $contadorArchivosIguales)"
             echo `mv $archivoEliminar $nombreCopia`
             archivoEliminar="$nombreCopia"
         fi
@@ -287,12 +289,11 @@ archivoParaBorrar="$1"
         then
             let contadorArchivosIguales=contadorArchivosIguales+1
             rutaArchivoUnico="$archivo"    # mejora
-            #archivosIguales="$archivosIguales$contadorArchivosIguales - $nombreArchivo $rutaArchivo;"
             archivosIguales="$archivosIguales$contadorArchivosIguales - $nombre $rutaArchivo;"
             arrayArchivos[$contadorArchivosIguales]="$archivo"
         fi
     done
-    echo "cantidad: $contadorArchivosIguales"
+    
     if [ "$contadorArchivosIguales" -eq 0 ];
     then
         echo "No existe el archivo en la papelera"
@@ -300,7 +301,8 @@ archivoParaBorrar="$1"
     else
         if [ "$contadorArchivosIguales" -eq 1 ];
         then   
-            tar -Pvf "$papelera" --delete "$rutaArchivoUnico" 1> /dev/null 
+            nombre=$(basename "$rutaArchivoUnico")
+            tar -Pvf "$papelera" --delete "$nombre" 1> /dev/null 
         else
             echo "$archivosIguales" | awk 'BEGIN{FS=";"} {for(i=1; i < NF; i++) print $i}'
             echo "¿Qué archivo desea borar?"
